requirements.txt
paho-mqtt==1.5.1
pymongo==3.11.2
click==7.1.2
XlsxWriter==1.3.7